Moved functions that query databse into a separate models file
